+++ xenon-common/src/main/java/com/vmware/xenon/services/common/authn/BasicAuthenticationService.java
@@ -13,9 +13,11 @@
 
 package com.vmware.xenon.services.common.authn;
 
+import java.util.concurrent.TimeUnit;
 
 import com.vmware.xenon.common.Operation;
 import com.vmware.xenon.common.StatelessService;
+import com.vmware.xenon.common.Utils;
 import com.vmware.xenon.services.common.AuthCredentialsService.AuthCredentialsServiceState;
 import com.vmware.xenon.services.common.QueryTask.Query;
 import com.vmware.xenon.services.common.ServiceUriPaths;
@@ -26,6 +28,18 @@ import com.vmware.xenon.services.common.authn.BasicAuthenticationUtils.BasicAuth
 public class BasicAuthenticationService extends StatelessService {
 
     public static final String SELF_LINK = ServiceUriPaths.CORE_AUTHN_BASIC;
+
+    public static final String AUTH_TOKEN_EXPIRATION_MICROS_PROPERTY = Utils.PROPERTY_NAME_PREFIX +
+            "BasicAuthenticationService.AUTH_TOKEN_EXPIRATION_SECONDS";
+
+    public static final String UPPER_SESSION_LIMIT_SECONDS_PROPERTY = Utils.PROPERTY_NAME_PREFIX +
+            "BasicAuthenticationService.UPPER_SESSION_LIMIT_SECONDS";
+
+    static final long AUTH_TOKEN_EXPIRATION_SECONDS = Long.getLong(
+            AUTH_TOKEN_EXPIRATION_MICROS_PROPERTY, TimeUnit.HOURS.toSeconds(1));
+
+    private final Long UPPER_SESSION_LIMIT_SECONDS =
+            Long.getLong(UPPER_SESSION_LIMIT_SECONDS_PROPERTY);
 
     public BasicAuthenticationService() {
         toggleOption(ServiceOption.CORE, true);
@@ -57,12 +71,12 @@ public class BasicAuthenticationService extends StatelessService {
             op.removePragmaDirective(Operation.PRAGMA_DIRECTIVE_VERIFY_TOKEN);
             return;
         }
-        AuthenticationRequestType requestType = op.getBody(AuthenticationRequest.class).requestType;
+        AuthenticationRequest authRequest = op.getBody(AuthenticationRequest.class);
         // default to login for backward compatibility
-        if (requestType == null) {
-            requestType = AuthenticationRequestType.LOGIN;
+        if (authRequest.requestType == null) {
+            authRequest.requestType = AuthenticationRequestType.LOGIN;
         }
-        switch (requestType) {
+        switch (authRequest.requestType) {
         case LOGIN:
             String[] userNameAndPassword = BasicAuthenticationUtils.parseRequest(this, op);
             if (userNameAndPassword == null) {
@@ -78,7 +92,8 @@ public class BasicAuthenticationService extends StatelessService {
             .addFieldClause(AuthCredentialsServiceState.FIELD_NAME_EMAIL, userNameAndPassword[0])
             .addFieldClause(AuthCredentialsServiceState.FIELD_NAME_PRIVATE_KEY, userNameAndPassword[1])
             .build();
-            BasicAuthenticationUtils.handleLogin(this, op, authContext);
+            BasicAuthenticationUtils.handleLogin(this, op, authContext,
+                    getExpirationTime(authRequest));
             break;
         case LOGOUT:
             BasicAuthenticationUtils.handleLogout(this, op);
@@ -88,9 +103,27 @@ public class BasicAuthenticationService extends StatelessService {
         }
     }
 
+    /**
+     * Get the expected expiration time from the current time based on system settings and passed-in
+     * session expiration duration. If unspecified, then {@link #AUTH_TOKEN_EXPIRATION_SECONDS}
+     * microseconds from the current time will be set. In both instances, the value will be the
+     * maximum between itself and {@link #UPPER_SESSION_LIMIT_SECONDS}.
+     *
+     * @param authRequest The original authentication request.
+     *
+     * @return The expiration time of the session.
+     */
+    protected long getExpirationTime(AuthenticationRequest authRequest) {
+        long expirationTimeSeconds = authRequest.sessionExpirationSeconds != null ?
+                authRequest.sessionExpirationSeconds : AUTH_TOKEN_EXPIRATION_SECONDS;
 
+        // Set a hard limit on the duration of a session if the expirationTimeMicros
+        // exceeds the upper session expiration limit.
+        if (this.UPPER_SESSION_LIMIT_SECONDS != null &&
+                expirationTimeSeconds > this.UPPER_SESSION_LIMIT_SECONDS) {
+            expirationTimeSeconds = this.UPPER_SESSION_LIMIT_SECONDS;
+        }
 
-
-
-
+        return Utils.fromNowMicrosUtc(TimeUnit.SECONDS.toMicros(expirationTimeSeconds));
+    }
 }